adding linux compatability

+++ RLFPS/Source/RLFPS/Private/HomingMod.cpp
@@ -32,7 +32,7 @@ void UHomingMod::OnSpawn_Implementation(ABullet* bullet)
 	TArray<AActor*> enemies;
 	UGameplayStatics::GetAllActorsOfClass(GetWorld(), bullet->targetClass, enemies);
 	AActor* closestPlayer = nullptr;
-	float distance = INT_MAX;
+	float distance = (float)INT_MAX;
 
 	//log all actors found for debugging
 	for (AActor* actor : enemies)
